Changes to be committed: 	new file:   .VSCodeCounter/2023-02-20_20-43-13/details.md 	new file:   .VSCodeCounter/2023-02-20_20-43-13/diff-details.md 	new file:   .VSCodeCounter/2023-02-20_20-43-13/diff.csv 	new file:   .VSCodeCounter/2023-02-20_20-43-13/diff.md 	new file:   .VSCodeCounter/2023-02-20_20-43-13/diff.txt 	new file:   .VSCodeCounter/2023-02-20_20-43-13/results.csv 	new file:   .VSCodeCounter/2023-02-20_20-43-13/results.json 	new file:   .VSCodeCounter/2023-02-20_20-43-13/results.md 	new file:   .VSCodeCounter/2023-02-20_20-43-13/results.txt 	modified:   lib/Academics/CGPA/CGPA.dart

+++ lib/Academics/CGPA/CGPA.dart
@@ -39,12 +39,12 @@ class _CGPAState extends State<CGPA> {
           GestureDetector(
             onTap: () {
               print("Gesture Detected");
-              int a = (int.parse(TEC_nowGPA.text.toString())) *
+              double a = (double.parse(TEC_nowGPA.text.toString())) *
                   (int.parse(TEC_nowSemCredit.text.toString()));
-              int b = (int.parse(TEC_oldCGPA.text.toString())) *
+              double b = (double.parse(TEC_oldCGPA.text.toString())) *
                   (int.parse(TEC_oldTotalCredits.text.toString()));
-              int c = (int.parse(TEC_nowSemCredit.text.toString())) +
-                  (int.parse(TEC_oldTotalCredits.text.toString()));
+              double c = (double.parse(TEC_nowSemCredit.text.toString())) +
+                  (double.parse(TEC_oldTotalCredits.text.toString()));
               showModalBottomSheet(
                   context: context,
                   backgroundColor: Colors.transparent,
